refactor: view로 전달하는 로직을 BlackjackGame으로 이동
