+++ src/main/java/blackjack/BlackjackGame.java
@@ -2,20 +2,34 @@ package blackjack;
 
 import blackjack.domain.BlackjackTable;
 import blackjack.domain.Command;
+import blackjack.domain.PlayerOutcome;
+import blackjack.domain.entry.Participant;
 import blackjack.domain.entry.Player;
+import blackjack.dto.CardCountingResult;
+import blackjack.dto.FirstTurnCards;
 import blackjack.dto.PlayerCardResult;
+import blackjack.dto.PlayerGameResult;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class BlackjackGame {
 
     public void start() {
         BlackjackTable blackjackTable = new BlackjackTable(InputView.inputNames());
-        OutputView.printFirstTurnCards(blackjackTable.getFirstTurnCards());
+        OutputView.printFirstTurnCards(toFirstTurnCards(blackjackTable.getParticipants()));
         hit(blackjackTable);
-        OutputView.printCardCountingResult(blackjackTable.getCardCountingResult());
-        OutputView.printGameResult(blackjackTable.getGameResults());
+        OutputView.printCardCountingResult(toCardCountingResult(blackjackTable.getParticipants()));
+        OutputView.printGameResult(toGameResults(blackjackTable.countGameResult()));
+    }
+
+    private List<FirstTurnCards> toFirstTurnCards(List<Participant> participants) {
+        return participants.stream()
+            .map(participant -> new FirstTurnCards(participant.getName(), participant.openCard()))
+            .collect(Collectors.toList());
     }
 
     private void hit(BlackjackTable blackjackTable) {
@@ -56,5 +70,33 @@ public class BlackjackGame {
 
     private PlayerCardResult toPlayerCard(Player player) {
         return new PlayerCardResult(player.getName(), player.getHoldCards().getCards());
+    }
+
+    private List<CardCountingResult> toCardCountingResult(List<Participant> participants) {
+        return participants.stream()
+            .map(this::toCardCountingResult)
+            .collect(Collectors.toList());
+    }
+
+    private CardCountingResult toCardCountingResult(Participant participant) {
+        return new CardCountingResult(
+            participant.getName(),
+            participant.getHoldCards().getCards(),
+            participant.countCards()
+        );
+    }
+
+    public List<PlayerGameResult> toGameResults(Map<PlayerOutcome, List<Player>> gameResults) {
+        return gameResults.keySet().stream()
+            .flatMap(outcome -> toPlayerGameResult(gameResults, outcome))
+            .collect(Collectors.toList());
+    }
+
+    private Stream<PlayerGameResult> toPlayerGameResult(
+        Map<PlayerOutcome, List<Player>> gameResult,
+        PlayerOutcome outcome
+    ) {
+        return gameResult.get(outcome).stream()
+            .map(player -> new PlayerGameResult(player.getName(), outcome));
     }
 }
